requirements.txt
cx-Oracle==8.3.0; sys_platform == 'linux'
cx-Oracle==8.3.0; sys_platform == 'win32'

pandas==2.2.3; sys_platform == 'linux'
pandas==2.2.3; sys_platform == 'win32'

SQLAlchemy==2.0.36; sys_platform == 'linux'
SQLAlchemy==2.0.36; sys_platform == 'win32'

python-dotenv==1.0.1; sys_platform == 'linux'
python-dotenv==1.0.1; sys_platform == 'win32'

tabulate ==0.9.0; sys_platform == 'linux'
tabulate ==0.9.0; sys_platform == 'win32'